Refactor code
- NotDeclarated -> NotDeclared
- traverse2 -> traverse
- Use Scala 3 parenthesis
